fix dll name

+++ build-stage2.sh
@@ -7,8 +7,8 @@ export DEBIAN_FRONTEND="noninteractive"
 
 DEFAULT_TARGET_LIST="aarch64-softmmu,alpha-softmmu,arm-softmmu,cris-softmmu,hppa-softmmu,i386-softmmu,lm32-softmmu,m68k-softmmu,microblaze-softmmu,microblazeel-softmmu,mips-softmmu,mips64-softmmu,mips64el-softmmu,mipsel-softmmu,moxie-softmmu,nios2-softmmu,or1k-softmmu,ppc-softmmu,ppc64-softmmu,riscv32-softmmu,riscv64-softmmu,rx-softmmu,s390x-softmmu,sh4-softmmu,sh4eb-softmmu,sparc-softmmu,sparc64-softmmu,tricore-softmmu,unicore32-softmmu,x86_64-softmmu,xtensa-softmmu,xtensaeb-softmmu"
 # default DLLs list is extracted from the installer provided at https://qemu.weilnetz.de/w64/
-DEFAULT_DLL_LIST="iconv,libatk-1.0-0,libbz2-1,libcairo-2,libcairo-gobject-2,libcurl-4,libeay32,libepoxy-0,libexpat-1,libffi-6,libfontconfig-1,libfreetype-6,libgdk-3-0,libgdk_pixbuf-2.0-0,libgio-2.0-0,libglib-2.0-0,libgmodule-2.0-0,libgmp-10,libgnutls-30,libgobject-2.0-0,libgtk-3-0,libharfbuzz-0,libhogweed-4,libidn2-0,libintl-8,libjpeg-8,liblzo2-2,libncursesw6,libnettle-6,libnghttp2-14,libp11-kit-0,libpango-1.0-0,libpangocairo-1.0-0,libpangoft2-1.0-0,libpangowin32-1.0-0,libpcre-1,libpixman-1-0,libpng16-16,libssh2-1,libstdc++-6,libtasn1-6,libunistring-2,libusb-1.0,libusbredirparser-1,libwinpthread-1,SDL2,ssleay32,zlib1"
-DEFAULT_DLL_LIST_GCC="libssp-0"
+DEFAULT_DLL_LIST="iconv,libatk-1.0-0,libbz2-1,libcairo-2,libcairo-gobject-2,libcurl-4,libeay32,libepoxy-0,libexpat-1,libffi-6,libfontconfig-1,libfreetype-6,libgdk-3-0,libgdk_pixbuf-2.0-0,libgio-2.0-0,libglib-2.0-0,libgmodule-2.0-0,libgmp-10,libgnutls-30,libgobject-2.0-0,libgtk-3-0,libharfbuzz-0,libhogweed-4,libidn2-0,libintl-8,libjpeg-8,liblzo2-2,libncursesw6,libnettle-6,libnghttp2-14,libp11-kit-0,libpango-1.0-0,libpangocairo-1.0-0,libpangoft2-1.0-0,libpangowin32-1.0-0,libpcre-1,libpixman-1-0,libpng16-16,libssh2-1,libtasn1-6,libunistring-2,libusb-1.0,libusbredirparser-1,libwinpthread-1,SDL2,ssleay32,zlib1"
+DEFAULT_DLL_LIST_GCC="libssp-0,libstdc++-6"
 DEFAULT_DLL_LIST_GCC_W32ONLY="libgcc_s_sjlj-1"
 
 SOURCE_BASE_DIR="${SOURCE_BASE_DIR:-${HOME}}"
@@ -27,9 +27,14 @@ MAKE_FLAGS="${MAKE_FLAGS:--j}" # note that -j might cause OOM (on a 32-core 128G
 # prepare sources
 mkdir -p "${SOURCE_BASE_DIR}"
 pushd "${SOURCE_BASE_DIR}"
-rm -rf ./qemu
 
-git clone "${SOURCE_GIT_URL}" qemu # the configure script will init the submodules, no need to do recursive here
+if [ -d "./qemu/.git" ]; then
+    echo "WARNING: qemu source exists, not cloning again"
+else
+    rm -rf ./qemu
+    git clone "${SOURCE_GIT_URL}" qemu # the configure script will init the submodules, no need to do recursive here
+fi
+
 pushd qemu
 git checkout "${SOURCE_GIT_REF}"
